+++ lib/src/acter_avatar.dart
@@ -4,24 +4,43 @@ import 'package:flutter/material.dart';
 /// ActerAvatar allows you to display the different types of Avatars in the
 /// acter.global app.
 class ActerAvatar extends StatefulWidget {
+  /// Set the display mode for this Avatar
   final DisplayMode mode;
+
+  /// The size this Avatar has
   final double? size;
-  final String? displayName;
+
+  /// the uniqueId of this object (e.g. full username or roomId)
+  /// used to calculate the Multiavatar in `DisplayMode.User`.
   final String uniqueId;
+
+  /// the display name they've chosen
+  final String? displayName;
+
+  /// a canonical uniqueName to use instead of the uniqueId in the tooltip, if given.
+  /// most commonly this is the canonical alias for a space/room rather than the roomID
+  final String? uniqueName;
+
+  /// If and how to display the tooltip
   final TooltipStyle tooltip;
+
+  /// The actual avatar (takes precedence)
   final ImageProvider<Object>? avatar;
+
+  /// Or alternatively a future that loads the avatar (show fallback until loaded)
   final Future<ImageProvider<Object>?>? avatarProviderFuture;
 
   ActerAvatar(
       {Key? key,
       this.displayName,
+      this.uniqueName,
       required this.uniqueId,
       required this.mode,
       this.tooltip = TooltipStyle.Combined,
       this.avatar,
       this.avatarProviderFuture,
       this.size})
-      : super(key: key);
+      : super(key: key ?? Key('avatar-$uniqueId-$size'));
 
   @override
   _ActerAvatar createState() => _ActerAvatar();
@@ -39,15 +58,18 @@ class _ActerAvatar extends State<ActerAvatar> {
 
   void setAvatar() async {
     if (widget.avatar != null) {
-      // hasAvatar == false
-      setState(() {
-        _avatar = widget.avatar;
-      });
+      if (mounted) {
+        setState(() {
+          _avatar = widget.avatar;
+        });
+      }
     } else if (widget.avatarProviderFuture != null) {
       final avatar = await widget.avatarProviderFuture;
-      setState(() {
-        _avatar = avatar;
-      });
+      if (mounted) {
+        setState(() {
+          _avatar = avatar;
+        });
+      }
     }
   }
 
@@ -61,7 +83,7 @@ class _ActerAvatar extends State<ActerAvatar> {
       case TooltipStyle.UniqueId:
         return Tooltip(message: widget.uniqueId, child: child);
       case TooltipStyle.Combined:
-        var message = widget.uniqueId;
+        var message = widget.uniqueName ?? widget.uniqueId;
         if (widget.displayName != null) {
           message = '${widget.displayName} (${widget.uniqueId})';
         }